// packages/Framework/src/UI/Data/List.types.ts
export enum ListImplementationType {
  flashlist = 'flashlist',
  flatlist = 'flatlist',
}
// ...
export type ListItem = {
  searchable: Record<string, string>;
  filterable: Record<string, string>;
  none: Record<string, string>;
};

/******************************************************************************************************************
 * Represent active filters applied to the list.
 *
 * @usage
 * ```ts
 * const filters: ListFilterMap = {
 *   material: new Set(['wood', 'cloth'])
 * }
 * ```
 ******************************************************************************************************************/
export type ListFilterMap = {
  [key: string]: Set<string>;
};
// ...
export type renderListItemFunc = (item: ListItem, index: number) => React.ReactNode;
// ...
export type ListProps = {
  dataArr: ListItem[];
  query: string;
  filterMap: ListFilterMap;
  renderItem: renderListItemFunc;
  listImplementationType?: ListImplementationType;
  emptyComponent?: React.ReactNode;
  style?: StyleProp<ViewStyle>;
};
// ...
export type ListType = React.FC<ListProps>;